Added transaction layer.
Contact is renamed to Nameaddr.

Leg now stores explicitly the own and peer's "Contact" address,
but without display name and tags. This is sent in/extracted from
INVITE messages only.

Leg's print and parse functions split into a make/take function, and
a print/parse one.

Leg.make_request no longer bothers with the branch generation, as this
is now the transport layer's task.

Implemented a bunch of Transaction classes with a TransactionManager.
The transactions are private to the TransactionManager, and it creates
them automatically for incoming and outgoing messages. The periodic
cleanup function is not used yet. The lower layer is a transport
object, while upwards the transactions call a report callable.

The test currently tries to use a bunch of components faked out.