Added CANCEL processing.
Sending a message modifies the user params.
This way the sender have access to all fields filled by the lower
layers, which is necessary to construct a CANCEL.

The sent INVITE is stored in Dialog for potential CANCELs.

The CANCEL is responded by the dialog if successfully found.
The Dialog also accepts To tag-less CANCELs.

Otherwise a stupid response is generated by the transaction layer.