Added Call, DirectMedia and Session classes.
The routing is now in Call.route(), overridden in the test.

The dial action now passes a ctx dict to do the routing by.

Report callbacks are not passed in constructors, but set with setters.

Leg is now a base class, functionality moved to SipLeg subclass.

SipLeg does not create its dialog, but gets in the constructor.

Method recved is renamed to process everywhere.

In the test Caller And Callee are now Leg subclasses, and placed
on the outer end of their Calls.