Hop structure.
The hop_addr is now hop everywhere. Transport cheats, and reverses
its fields during transmission.

The DialogManager can access the Transport's get_hop method to
determine the Hop of a Uri.

Routing is done on a copy of the src_ctx.

The Dialog's setup_{incoming,outgoing} is no longer called by the
Switch, but by the Leg (outgoing), and the Dialog itself (incoming).
The take_request needed a little change for this.