Added Authority class.
Digest authorization works on a basic level. Dialog retries, and
the Switch authenticates.

Dialog doesn't fetch the credentials, but it is set on it.

Added incoming CSeq check in Dialog.

Reporting responses also include the sent request.

Added WwwAuth and Auth header fields to format.py.

Added make_ack function fo consistency.

The report argument is no longer a Transaction constructor argument,
because the signature of this event hadler is different for requests and
responses.

Resending the INVITE is horrible.

Reporting responses for the resent INVITE does not work.