Authorization works properly.
We don't try authorization again if the request already contained
an Authorization header.

Dialog no longer gets a response callback for requests, it uses the
plain callback in all cases.

The user_params is saved in the extended params of the requests,
and is used when retrying an unauthorized request.

In-dialog checks cleaned up.

In-dialog authorization works.

The test authorizes both SIP legs.