Session handling improvements in Leg.
Extended InviteState.

Added another implementation of Session. This may replace the
one in Call.

Added DIALING_OUT_ANSWERED state for waiting for local session
answer if it needs to be sent in the ACK.

The action dict-s no longer have sdp key, but separate offer
and answer for the sake of strictness.

Added session type actions for session-only operations.

Added status code 183 handling.

Prepared the code to handle non-session SDP-s in the future.

Added EmptySdp base class to be used for session queries and
rejections, soon.

Updated test.