Msgp/Mp modifications.
Msgp allows request_handler-s set for each stream in add_stream.

An incoming stream is unconfirmed until an add_stream, but still
can be used to send a response before removing it. Such removals are silent,
no unsolicited nope-s are sent.

Received nope-s are reported through the request_handler-s.

Context is created without parameters, and later configured
with modifications.

Context has a dedicated request_handler for modify and delete,
the latter notifies the ContextManager.

Context records the owner_addr in each modification.

In the Mp, requests are handled by their target.

And exceptions result in an error response.

Controller no longer needs to store request_handlers, as it is now done
in the Msgp.