Massive routing refactor.
Routing is now performed in the Routing(s), created by Call. Routing-s
can send routing events up to Call. Leg-s can call media related methods
of Call directly. Routing ends with the anchor event, which is supposed
to tear down every outgoing branches but one, and then a single list of
Leg-s are stored in the Call directly.

Finish events signal the end of life from Leg-s and Routing-s.

Session management is moved from Call to Leg. Leg-s can ask for media
addresses from Call, which in turn can use the Mgc for it.

Refresh calls from Leg-s to Call grab meida leg infos from the Leg-s,
and updates the media channel contexts. This is probably not a good
approach, and will be redesigned soon.

MediaLeg-s are now quite dumb, and only store a dict of values, but
this may change.

Planner can now raise Exceptions if passed to resume.

The CalleeLeg is now a bit hackish because of the media handling
misdesign.

Added TestRouting and TestCall helper classes to the test.

Test call from phone to 555 works.