Using the Python logging module.
Each module has it own logger.

Authority no longer gets a realm argument, instead uses the get_realm
method that can be overloaded, and by default returns the request's To
domain.

Routing.route_call now called properly with error handling.

Dialog.setup_outgoing now gets an explicit RURI argument, which may differ
from the To URI for forwarded calls. From and To are passed as Nameaddr-s.
The RURI is used for determining the hop, as the To may be symbolic.

SipLeg calls it this way.

In the test constants are now uppercase. We use wymbolic From and To
domain names, which are distinct from the RURI-s and Contact addresses.

The test media played is shorter.

TestAuthority is more detailed in logging.

Switch no longer gets a realm argument, but a list of domains it is
responsible for.

Logging is also configured here.

UdpTransport.get_hop handles correctly the unspecified port case,
defaulting to 5060.

The exchandler now handles exceptions during the serialization of
local variables.