Authorization refactorings.
Authority now also gets to decide if a certain authname is
permitted to use a certain address.

authenticate() is renamed to authorize(), and identify() is
renamed to get_remote_credentials().

Added SimpleAuthority subclass that authorized in two steps, first
without looking at the authorization infos, then looking up the
ha1 and the list of allowed URI-s by authname. The test uses this.

Call anchored on a reject.

RecordManager has a chance to reject the creation of a record by
is_known_record().

In the test the CALLER and CALLEE creds renamed to A and B.

TestAuthority is now a subclass of SimpleAuthority, and expects the
list of allowed URI-s for local credentials.