Added WeakGeneratorMethod class.
This is a subclass of WeakMethod, which now has a make_selfarg
method to create the self argument of the actual call. This
class overloads it to pass a proxy object for self.

Call and Leg now uses this class.

Planner expects such objects, does not dissects the received bound
method.