Routing improvements.
Added WeakMethod.bind_front() method.

Basic process method moved to the base Routing class. Added
helper methods to help overloading.

Action to hang up an unanswered call is now 'cancel'.

Removed anchor action, instead added Leg.get_further_legs(). This way
anchoring is only reported to the owner of a Routing, and if it happens
to be an internal leg, it will store the result in itself, to be
retrieved by this new method.