DialOutLeg works.
SipLeg didn't fill the URI of the ctx.

Routing is now done by the request URI, not the To URI, which seems to be
really useless, as only the request URI may be rewritten by proxies,
therefore must be the basis of further routing decisions.

SwitchB now spawns a dialout leg, which spawns the 3 test legs.