Added Bridge class.
Refactored Dial*Leg-s to a Bridge and BridgeLeg classes. The BridgeLeg-s
hold strong reference to the Bridge only. This makes the finishing
a bit tricky, as the Bridge-Routing-BridgeLeg is a reference cycle
for a while. The Bridge initiates the finish of the BridgeLeg-s
when necessary.

Routable is now officially Loggable.

Switch.make_outgoing_leg takes a call argument, too.