100rel now works with the Snom, too.
Fixed name of the Require header.

InviteState remembers if a session was sent in a rpr.

InviteState can queue outgoing invite responses when necessary,
and send them after responding to a PRACK.

Merged the invite_outgoing_response* methods, because one was only
useful for returning if sending a session was possible.

Final invite responses were also sent as rpr.

Incoming PRACK is properly checked, and stops the rpr retransmission.

Simplified sending the 100 trying.

rpr-s in InviteServer are now transmitted like requests until
stopped.