Msgp fix for concluding handshakes.
The last handshake message is now sent by the Msgp layer.

Users not sending further handshake messages must call done_handshake
to finish it locally.

Users must use send_handshake for outgoing messages, and omit
the response handler in the last step to let the Msgp layer
catch the last answer.

Handshakes are again in a separate container for clarity.