Plan is a separate object from the planned.
Planned being a superclass was a bad idea, since even if the toplevel
generator gets a weak reference to self, any further subgenerator
gets a strong reference, so suspending a plan would still have strong
references, making the planned object indestructible. So now Plan
is a subobject, having a weak reference to the planned one, which
can be destroyed, and then explicitly aborting the Plan, terminating
the suspended generators.

Plan no longer gets a finished callback, but calls a method,
to be overloaded.

sdp is now an optional argument of InviteState.outgoing.

Removed unused code from siplibtest.

The test works.