Plan, resume, and instaplug refactors.
Removed sync option from Plug, instead created a subclass InstaPlug
with the same functionality, and a similar helper method in Slot.

Renamed inplug and unplug methods of slot to plug_in and plug_out for
clarity, removed unplug_all, because it was the only method to be called
from the outside.

Allow Slot.zap to be called with args to simplify EventSlot's operation.

Improved Slot.zap so removed plugs are not zapped accidentally.

Plan now uses instaplugs, and stores the resume value in the Plan
instead of passing it to resume.

resolver uses a socket timeout of 5 seconds.

MGW Context now properly stores its forward plugs.