MGW affinity is stored in sessions.
This is the first step to eliminate Call.

MGW sid can be selected by channel type and affinity.

Media address can be selected by sid.

Media leg can be created by type and sid.

Cleaned up a bit SipParty.allocate_local_media.
When allocating a payload type for a format where the remote
choice is present, try choosing that one. This can save a
renumbering in the MGW.