Spawning parties uses the dst of the dial action.
Dialog now generates the local tag immediately, as this is used
to identify SipEndpoint.

Ground now allows forwarding a dial action without linked leg, in
this case it extracts the dst field of the dial action to create
and link the new party, or creates a routing if it is missing.

A call_info dict is set to Party-s instead of call_oid and path.
This is also used to number routings within a call, which are otherwise
identity-less.

Ground.setup_leg is now make_leg.

Party-s now have oids in the form of type=identity (or routing=number).

Removed Party.set_path, added {set,get}_call_info and identify.

Party.dial is now just a simple function that updates a dial action,
and forwards it on a new leg. The argument order is changed.

SimpleBridge is now totally commented out.

An outgoing SipEndpoint is now set up by its dst dict, not the ctx.

An incoming SipEndpoint now sets up a src field in the dial action,
not the ctx.

Switch now creates a default routing, but that rejects everything.

The ctx starts as empty when the first routing is run.

TestLine identity is also set up in the dst dict.

Test routings are updated to use the new logic.