RPR reverts and cleanups.
Brought back the unreliable provisional message sending in InviteServerState.

Brought back the PRACK offer handling, and also fixed it in SipEndpoint,
as we may need to reject pending offers when PRACK offer arrives.

RPR is now either required from peers, or not even offered. This way
we can be sure if it will be used, or not, and don't have to deal with
the hopeless cases when one side does and the other doesn't.

We currently use RPR on all INVITE-s.

Unreliable provisional responses abort processing if RPR was required.

The old pending_sdp variable is now provisional_sdp, and is only
present in InviteServerState.

First response without offer is not sent if RPR are required.

Fixed rpr without answer case that was always sent reliably even
without RPR support.

Added exceptions when an outgoing operation passed neither message nor sdp.