Further RPR cleanups and fixes.
InviteState-s use instaplugs for sending messages, because otherwise
the ordering of messages can change, and a subsequent offer can be sent
earlier in an UPDATE, than the previous answer in a PRACK.

Added is_answer flag to {invite,update}_outgoing methods, so that
they can be checked for safety in the transaction-specific code.

update_outgoing ignored the passed msg argument and created its own.

It now also handles session rejection.

update_incoming didn't clear the update helper object after receiving
the response, and didn't store the request after receiving it, and didn't
return the received msg.

Also added update_incoming_sdp helper method.

process_outgoing_session also return the is_answer flag, even if it's
trivial to get.

And doesn't call realize_local_media and builds SDP for session rejections.

process_outgoing_session is called at the beginning of the do method again,
not per case.

Added InviteState will_session_prack method. This is used to decide if
the outgoing session will generate a PRACK, and this way we no longer pass
a None msg argument to invite_outgoing, but create the messages ourselves.

When dialing in, an outgoing session may not only be ripped to an
UPDATE, but if a PRACK offer is pending, then we rip it to the PRACK response,
too.

Rejecting incoming INVITE-s with wrong RPR support flags is now done in
InviteServerState.

If an incoming UPDATE was bogus, we no longer report bogus sessions.

Added KEEP state to InviteState-s instead of using None to keep the state.

Added ABORT state to make the difference to FINISH more pronounced.

RELIABLE_NOANSWER is renamed to RELIABLE_PREANSWER, and RELIABLE_EMPTY to
RELIABLE_POSTANSWER.

Removed the ambiguous abort and ignore methods, because they could only be
used in place of recv, not send, yet it was used that way at many places.
Now the new ABORT and KEEP states must be used explicitly.

Using make_rack method instead of making the whole PRACK request, because
now it is passed from SipEndpoint (instead of None msg).

Similarly the explicit ACK request is also passed in, not constructed locally.
We can still ACK/PRACK implicitly after receiving final/reliable responses.

Instead of raising Exceptions we always transition to the ABORT state.

In InsiteServerState some cases are now separated to make it clear when the
outgoing SDP is just repeated.

PRACK offer handling is back, if it wasn't in the previous commit.

In the test Caller also forces an early UPDATE.