Took care of MGW responses to dead things.
The MGC instead of using a WeakValueDictionary, now uses a dict
that stores weak reference to things. This way a response can see
if the thing existed, and the final response removes the entry.

Thing-s no longer assemble their own response_tag.

And by the way response_tag was renamed to origin to be more
symmetric to target in requests.

Rephrased some MSGP log messages.