Lame Alert-Info support.
SipEndpoint can pick up an outgoing Alert-Info header from dst.

An in-dialog session answer can be sent out on an active UPDATE,
creating a re-INVITE in this case is bad.

Made the test Caller a bit more precise in the session handling,
so it won't get into an infinite offer/answer loop.