Now we can handle multiple registrations properly.
The contacts are now stored per call-id, and separate cseq for each,
so multiple nondialogs can be maintained.

Added Uri.contains method, and the scheme-less version must use
None as the scheme, not "any", even it it is how it's printed.

Added Record.add_static_contact for a cleaner solution.

We now explicitly reject REGISTER requests with multiple Contact
values, because the per callid solution can't handle it.