More session handling improvements.
SessionState now returns a result code instead of just raising
exceptions. Even session queries can be ignored in some cases.

Added SipEndpoint.try_update to send a session with UPDATE if we can.

Added SipEndpoint.process_update to process incoming UPDATE-s, because
this is exactly the same in all states.

In DIALING_IN state a rejected session didn't rejected the INVITE.

We can send an answer in the ACK of a re-INVITE.

Added lost method check in SipEndpoint.process for initial INVITE-s.

Removed UpdateHelper.update_{incoming,outgoing}_auto methods.

SessionHelper.process_incoming_sdp can ignore incoming sessions
if SessionState had a problem with it.

SessionHelper.process_outgoing_session can ignore or reject outgoing
sessions if SessionState had a problem with it.

InviteState objects now explicitly know their direction. This is
necessary because an outgoing answer in a re-INVITE can be sent
in both a response and an ACK.

Fixed the test importing the renamed AlertInfo, and not alerting now.