Accounts properly split into local and remote ones.
Authority is now a base class for Account.

Account is now a base class for LocalAccount and RemoteAccount.

LocalAccount can check requests, and generate challenges, and are
stored by authname.

RemoteAccount can check responses and provide authentications, and are stored
by the remote AoR they're used for.

Realm is no longer deduced from the request, but specified for the account.

We no longer check the incoming request URI, as it is useless for
the digest check, instead using the URI in the digest header. Similarly
for outgoing authentications we use the To URI for digests, the server
must accept it this way.

Removed the stale hack from the Auth class, instead we check the staledness
in require_auth.

Other modules no longer ask for the required authentication headers, only
for the RemoteAccount object, and ask it directly.

Switch now separately calls check_auth and require_auth.

TestSwitch now also expects the realm name in the constructor, and uses
it for creating local accounts.

The test switches create the remote account specifically to each other's
domain.