Made Registrar updates RFC compliant.
Now the call_id and the cseq are stored for each contact, and are
checked for each incoming request. If any of them is out of order,
then the request is rejected. Otherwise thay're all updated.

Gave up on using the regular Dialog-s for registrations, and
merged methods to simplify code.