Added RedialBridge to generate artificial ringback tone.
This was factored out of TestLine.

Fixed the naming of Status.SERVER_INTERNAL_ERROR everywhere.

Transfer usage of Sip.{request,response} didn't use keys for header fields.

IUC logs the cause why it may not finish.

Some return values and parse_sdp calls were also wrong.

And unresponded_invite was not cleared after sending the response.

Ground.blind_transfer now spawns a RedialBridge as the next party.

DtmfInjector was rewritten to be more flexible, and don't block packets
when the stream is restarted with older timestamps. Now it only
outputs the DTMF packet when the next packet if filtered, so it does
not need to know the packet durations.

Added BlindTransferringEndpoint to test blind transfers, and replaced
the reringing endpoint with it (reringing is now forbidden).

Fixed AckClient transaction retransmitting the message due not changing
to WAITING state.