Call parking works using blind transfer and pickup.
Remote accounts are stored by realm, not URI, because storing by URI was
weird, as resolving hostnames is awkward, adding both resolved and
unresolved entries is even more awkward, and both are wrong, because
proxy authentication will only work if we make a distinction by realm.

Authority.check_digest no longer checks the realm, it is done
by its caller.

The local realm is stored in the AccountManager, not per LocalAccount-s.

Added AccountManager.provide_auth, as this mathod can also look up the
RemoteAccount by the realm, no need to return it.

SipEndpoint swallows a late ACK, as the Snom can transfer a call when
the ACK for a re-INVITE is still on its way.

IUC resets its state when the transaction is complete, not when the next
one starts. It has a new start method for that.

Leg couldn't reject wrong ground sessions, because it would also drop
its rejection. Now we forward those rejections directly.

SessionNegotiatorBridge starts with sendrecv direction on all channels,
otherwise a peer put on hold before the transfer would continue to
stay on hold afterwards, as the Snom does this to Caller.

Added an extension to a Line-wrapped Callee, but not using it right now.

Added explicit realm constants.

Allow TestEndpoint to be idle without media legs.

RingingEndpoint now properly finishes the session negotiation before
going ringing, and goes idle to keep sessions happy until hangup.

CalleeEndpoint's UPDATE-forcing extra sessions are commented out, as
it confuses the more complex scenarios.

The Snom can now do a transfer to BLF button, and pickup from ringing BLF
button, but the Park+Orbit functionality doesn't seem to work yet.