Rewrote MGW to use media things with slot based forwarding.
The MGW only creates things, and instead of contexts it just
links their slots together.

The create/modify/delete operations now work on things.

Create is called for all things quite soon, we no longer want
to optimize the first modify out, as it is error prone, and we
don't care about performance that much.

In the MGC, MediaLeg is now only a namedtuple to store some info
about a slot of a media thing, and has no useful methods at all.

Party now maintains an array of media_things.

PassMediaLeg is now RecordMediaThing.

NetMediaLeg is now RtpMediaThing.