Attended transfer now also uses the spawning approach.
One leg is hung up explicitly, while the other spawns a
SessionNegotiatorBridge, which later dials using the new grab dst type,
which finally links the other leg instead of creating a new one.

Noticed that the Snom generates invalid unbracketed URI with headers in
the Refer-To header, unless configured otherwise.

Reason causes are made sure to be strings, as it is a param value,
and will be parsed to string as well.

Leg.get_channel_count renamed to get_potential_channel_count, and it
now considers sessions in both directions, and even the pending sessions,
because accepting a session won't trigger a media leg linking.

Anchoring and unanchoring legs also updates the media leg linkage.

Renamed Ground.find_similar_media_leg to find_backing_media_leg.

Ground.spawn_transfer method merged into transfer_leg.

Transferring legs are now always hung up explicitly, even it is not
really necessary for deflect, as it's an incoming 302 response, but
doesn't hurt.

CalleEndpoint waits for 3 seconds before answering, because the edited
ringing.wav begins with 2 seconds of silence.

The code does not work right now, because Party unanchores its legs
before collapsing, which seems to be a silly idea, and will be
changed in a moment. But now the collapsing session negotiator
doesn't link the two sides of media after an attended transfer.