Collapsing legs now requires anchoring.
Until now they had to be unanchored, which seemed to be a bad idea.

Routing.anchor_outgoing_leg now also calls Bridge's version, just makes
sure the queues are not flushed there, because it must be done in the
collapsing step here.

Attended and blind transfers seem to work again.

Lines starting # with '#' will be ignored, and an empty message aborts
the commit. # On branch master # Changes to be committed: #	modified:
ground.py #	modified:   party.py #