HttpLikeStream now uses RFC5626 CRLF keepalives.
Removed unused slots.

CRLF handling is quite awkward, as one can never be sure after receiving
one that if it's a PONG, or the first half of a PING. So we cancel the
timeout for all CRLF, but treat every second one (in a consecutive
series) as a PING that needs to be PONG-ed.