MsgpStream again keeps track of the last received seq.
This makes it possible to ignore duplicates after a reconnection, and
also sending this info to the peer with the handshake.

The @hello message uses a target token to express the encoding. Currently
unused.

The @bello message uses a target token that ACK-s the received messages.

Fixed obsolete reference to SipLikeMessage.