Cleanups for digest authentication.
Renamed WwwAuth and Auth classes for clarity.

Passed algorithm into Authority.compute_digest so it's checked
at one place only, and case-insensitively (this seems to be right, as
the ABNF literals are all case-insensitive, and the Cisdo uses
the lowercase form).

Fixed typo when accessing self.nonce.

Added Cisco number and record, unfortunately using the resolved
IP address as domain as there seems to be no way to set a domain
in the XML config.

Made UDP message parsing a bit more robust, as the Cisco sent in
some leading garbage for a reassembled fragmented UDP packet.