Not requiring rpr anymore.
The Cisco doesn't support it.

Now we only deny it if the invite has no SDP, because that would lead
to a screwed up scenario. Otherwise we support it, but the peer may
choose not to.

Improved Bridge.do_leg to allow ring actions with session be forwarded
to the incoming leg if anchored and no previous ringing was sent.