Working Snom presence notifications.
The empty URI scheme no longer printed as any:.

State-s and EventSource-s now use an uri parameter, not entity, because
this term is inconsistent in different formats. The uri is used in its
canonical AoR form.

DialogFormatter uses this uri as the dialog-info entity, but PresenceFormatter
as contact, and uses the bare domain with pres: scheme as presence entity.

Added contact field to Presence Formatter because the Snom example
XML parsing checks for it.

Removed whitespace around XML text nodes because it is not stripped by
the Snom's parser.

Fixed PhoneLocalState.get_state not returning a flat dict.

Added pretty printing to EventKey.